refactor(runner): move FailedStatus from ClaudeRunner to ProcessRunner
Removes the hard dependency ProcessRunner had on ClaudeRunner::FailedStatus,
allowing ProcessRunner to be used independently. Adds a backwards-compat
alias in ClaudeRunner so existing references continue to work.
